List tile + colors working
